Fix audit logging in RoleService and refactor into modular structure #37
